Auto-mail when question is added to group + Created pitch outline (and TODO)
Auto-mail whenever a question (pre-existing or new) is added to a
group.
As yet it does not give a link to the question because drawcery is not
currently hosted anywhere.

Added an outline for the initial pitch (PitchPresentation.md) to
outline the functions given.
This has an associated TODO list, which will be the focus of the
upcoming week.